feat(neon_lints): add avoid_exports lint rule
