+++ packages/neon/neon_files/lib/src/blocs/browser.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:built_collection/built_collection.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:neon_files/src/options.dart';
 import 'package:neon_framework/blocs.dart';
 import 'package:neon_framework/models.dart';
@@ -25,7 +24,6 @@ enum FilesBrowserMode {
 }
 
 sealed class FilesBrowserBloc implements InteractiveBloc {
-  @internal
   factory FilesBrowserBloc({
     required FilesOptions options,
     required Account account,
