+++ packages/neon/neon_files/lib/src/blocs/files.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:built_collection/built_collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:neon_files/l10n/localizations.dart';
 import 'package:neon_files/src/blocs/browser.dart';
 import 'package:neon_files/src/options.dart';
@@ -23,7 +22,6 @@ import 'package:timezone/timezone.dart' as tz;
 import 'package:universal_io/io.dart';
 
 sealed class FilesBloc implements InteractiveBloc {
-  @internal
   factory FilesBloc({
     required FilesOptions options,
     required Account account,
