+++ packages/neon/neon_news/lib/src/blocs/article.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:neon_framework/blocs.dart';
 import 'package:neon_framework/models.dart';
 import 'package:neon_news/src/blocs/articles.dart';
@@ -9,7 +8,6 @@ import 'package:nextcloud/news.dart' as news;
 import 'package:rxdart/rxdart.dart';
 
 sealed class NewsArticleBloc implements InteractiveBloc {
-  @internal
   factory NewsArticleBloc({
     required NewsArticlesBloc articlesBloc,
     required Account account,
