+++ packages/neon/neon_news/lib/src/blocs/articles.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:built_collection/built_collection.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:neon_framework/blocs.dart';
 import 'package:neon_framework/models.dart';
 import 'package:neon_framework/utils.dart';
@@ -23,7 +22,6 @@ enum ListType {
 }
 
 sealed class NewsArticlesBloc implements InteractiveBloc {
-  @internal
   factory NewsArticlesBloc({
     required NewsBloc newsBloc,
     required NewsOptions options,
