+++ packages/neon/neon_news/lib/src/blocs/news.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:built_collection/built_collection.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:neon_framework/blocs.dart';
 import 'package:neon_framework/models.dart';
 import 'package:neon_framework/utils.dart';
@@ -12,7 +11,6 @@ import 'package:nextcloud/news.dart' as news;
 import 'package:rxdart/rxdart.dart';
 
 sealed class NewsBloc implements InteractiveBloc {
-  @internal
   factory NewsBloc({
     required NewsOptions options,
     required Account account,
