+++ packages/neon/neon_notes/lib/src/blocs/notes.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:built_collection/built_collection.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:neon_framework/blocs.dart';
 import 'package:neon_framework/models.dart';
 import 'package:neon_framework/utils.dart';
@@ -11,7 +10,6 @@ import 'package:nextcloud/notes.dart' as $notes;
 import 'package:rxdart/rxdart.dart';
 
 sealed class NotesBloc implements InteractiveBloc {
-  @internal
   factory NotesBloc({
     required NotesOptions options,
     required Account account,
