+++ packages/neon/neon_notes/pubspec.yaml
@@ -18,7 +18,6 @@ dependencies:
   http: ^1.2.1
   intl: ^0.18.0
   logging: ^1.0.0
-  meta: ^1.0.0
   neon_framework:
     git:
       url: https://github.com/nextcloud/neon
