+++ packages/neon/neon_notifications/lib/src/blocs/notifications.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:built_collection/built_collection.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:neon_framework/blocs.dart';
 import 'package:neon_framework/models.dart';
 import 'package:neon_framework/utils.dart';
@@ -11,7 +10,6 @@ import 'package:nextcloud/notifications.dart' as $notifications;
 import 'package:rxdart/rxdart.dart';
 
 sealed class NotificationsBloc implements NotificationsBlocInterface, InteractiveBloc {
-  @internal
   factory NotificationsBloc({
     required NotificationsOptions options,
     required Account account,
