+++ packages/neon/neon_notifications/pubspec.yaml
@@ -16,7 +16,6 @@ dependencies:
   go_router: ^13.0.0
   intl: ^0.18.0
   logging: ^1.0.0
-  meta: ^1.0.0
   neon_framework:
     git:
       url: https://github.com/nextcloud/neon
