+++ packages/neon_lints/example/lib/main.dart
@@ -1,0 +1,2 @@
+// expect_lint: avoid_exports
+export 'src/some_library.dart' hide InternalClass;
