+++ packages/neon_lints/example/lib/src/some_library.dart
@@ -1,0 +1,7 @@
+import 'package:meta/meta.dart';
+
+// expect_lint: avoid_exports
+@internal
+class InternalClass {}
+
+class ExportedClass {}
