+++ packages/neon_lints/lib/neon_lints.dart
@@ -2,6 +2,7 @@ import 'package:custom_lint_builder/custom_lint_builder.dart';
 import 'package:neon_lints/src/avoid_dart_html.dart';
 import 'package:neon_lints/src/avoid_dart_io.dart';
 import 'package:neon_lints/src/avoid_debug_print.dart';
+import 'package:neon_lints/src/avoid_exports.dart';
 
 /// Registers the neon lints plugin.
 ///
@@ -16,5 +17,6 @@ class _NeonLintsPlugin extends PluginBase {
         AvoidDebugPrint(),
         AvoidDartHTML(),
         AvoidDartIO(),
+        AvoidExports(),
       ];
 }
